move error checks to separate file
