+++ stac_fastapi/mongo/stac_fastapi/mongo/transactions.py
@@ -5,7 +5,6 @@ from datetime import datetime
 
 import attr
 from stac_pydantic.shared import DATETIME_RFC339
-from pymongo import MongoClient
 
 from stac_fastapi.extensions.third_party.bulk_transactions import (
     BaseBulkTransactionsClient,
@@ -15,61 +14,11 @@ from stac_fastapi.mongo.config import MongoSettings
 from stac_fastapi.mongo.serializers import ItemSerializer
 from stac_fastapi.mongo.session import Session
 from stac_fastapi.types import stac as stac_types
+from stac_fastapi.mongo.error_checks import ErrorChecks
 from stac_fastapi.types.core import BaseTransactionsClient
-from stac_fastapi.types.errors import ConflictError, ForeignKeyError, NotFoundError
 from stac_fastapi.types.links import CollectionLinks, ItemLinks
 
 logger = logging.getLogger(__name__)
-
-@attr.s
-class ErrorChecks():
-    session: Session = attr.ib(default=Session)
-    client: MongoClient = attr.ib(default=None)
-
-    def _check_collection_foreign_key(self, model: stac_types.Item):
-        if not self.client.stac.stac_collection.count_documents(
-            {"id": model["collection"]}, limit=1, session=self.session
-        ):
-            raise ForeignKeyError(
-                f"Collection {model['collection']} does not exist"
-            )
-
-    def _check_collection_conflict(self, model):
-        if self.client.stac.stac_collection.count_documents(
-            {"id": model["id"]}, limit=1, session=self.session
-        ):
-            raise ConflictError(f"Collection {model['id']} already exists")
-
-    def _check_collection_not_found(self, collection_id):
-        if (
-            self.client.stac.stac_collection.count_documents(
-                {"id": collection_id}, session=self.session
-            )
-            == 0
-        ):
-            raise NotFoundError(f"Collection {collection_id} not found")
-
-    def _check_item_conflict(self, model: stac_types.Item):
-        if self.client.stac.stac_item.count_documents(
-            {"id": model["id"], "collection": model["collection"]},
-            limit=1,
-            session=self.session,
-        ):
-            raise ConflictError(
-                f"Item {model['id']} in collection {model['collection']} already exists"
-            )
-
-    def _check_item_not_found(self, item_id, collection_id):
-        if (
-            self.client.stac.stac_item.count_documents(
-                {"id": item_id, "collection": collection_id},
-                session=self.session,
-            )
-            == 0
-        ):
-            raise NotFoundError(
-                f"Item {item_id} in collection {collection_id} not found"
-            )
 
 
 @attr.s
